+++ routes/main.php
@@ -22,15 +22,13 @@ use Adshares\Adserver\Http\Controllers\InfoController;
 use Adshares\Adserver\Http\Controllers\Manager\CampaignsController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/',
-    function () {
-        return '';
-    })->name('login');
-
-# API INFO
+Route::get('/', [InfoController::class, 'info'])
+    ->name('login');
 Route::get('/info', [InfoController::class, 'info']);
-Route::get('/info.json', [InfoController::class, 'info'])->name('app.infoEndpoint');
-Route::get('/upload-preview/{type}/{name}', [CampaignsController::class, 'uploadPreview'])->name('app.campaigns.upload_preview');
+Route::get('/info.json', [InfoController::class, 'info'])
+    ->name('app.infoEndpoint');
+Route::get('/upload-preview/{type}/{name}', [CampaignsController::class, 'uploadPreview'])
+    ->name('app.campaigns.upload_preview');
 
 Route::get('/policies/privacy.html', [InfoController::class, 'privacyPolicy']);
 Route::get('/policies/terms.html', [InfoController::class, 'terms']);