+++ routes/manager.php
@@ -29,31 +29,45 @@ use Adshares\Adserver\Http\Controllers\Manager\StatsController;
 use Adshares\Adserver\Http\Controllers\Manager\WalletController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('campaigns', [CampaignsController::class, 'browse'])->name('app.campaigns.browse');
-Route::get('campaigns/count', [CampaignsController::class, 'count'])->name('app.campaigns.count');
-Route::get('campaigns/{campaign_id}', [CampaignsController::class, 'read'])->name('app.campaigns.read');
-Route::post('campaigns', [CampaignsController::class, 'add'])->name('app.campaigns.add');
-Route::patch('campaigns/{campaign_id}', [CampaignsController::class, 'edit'])->name('app.campaigns.edit');
+Route::get('campaigns', [CampaignsController::class, 'browse'])
+    ->name('app.campaigns.browse');
+Route::get('campaigns/count', [CampaignsController::class, 'count'])
+    ->name('app.campaigns.count');
+Route::get('campaigns/{campaign_id}', [CampaignsController::class, 'read'])
+    ->name('app.campaigns.read');
+Route::post('campaigns', [CampaignsController::class, 'add'])
+    ->name('app.campaigns.add');
+Route::patch('campaigns/{campaign_id}', [CampaignsController::class, 'edit'])
+    ->name('app.campaigns.edit');
 Route::put('campaigns/{campaign}/status', [CampaignsController::class, 'changeStatus'])
     ->name('app.campaigns.change_status');
 Route::put('campaigns/{campaign_id}/banner/{banner_id}/status', [CampaignsController::class, 'changeBannerStatus'])
     ->name('app.campaigns.change_banner_status');
-Route::delete('campaigns/{campaign_id}', [CampaignsController::class, 'delete'])->name('app.campaigns.delete');
+Route::delete('campaigns/{campaign_id}', [CampaignsController::class, 'delete'])
+    ->name('app.campaigns.delete');
 
-Route::post('campaigns/banner', [CampaignsController::class, 'upload'])->name('app.campaigns.upload');
+Route::post('campaigns/banner', [CampaignsController::class, 'upload'])
+    ->name('app.campaigns.upload');
 
 Route::post('campaigns/{campaign_id}/classify', [CampaignsController::class, 'classify'])
     ->name('app.campaigns.classify');
 Route::delete('campaigns/{campaign_id}/classify', [CampaignsController::class, 'disableClassify'])
     ->name('app.campaigns.disable_classify');
 
-Route::post('sites', [SitesController::class, 'create'])->name('app.sites.add');
-Route::get('sites/sizes/{site_id?}', [SitesController::class, 'readSitesSizes'])->name('app.sites.sizes');
-Route::get('sites/{site}', [SitesController::class, 'read'])->name('app.sites.read');
-Route::patch('sites/{site}', [SitesController::class, 'update'])->name('app.sites.edit');
-Route::delete('sites/{site}', [SitesController::class, 'delete'])->name('app.sites.delete');
-Route::get('sites', [SitesController::class, 'list'])->name('app.sites.browse');
-Route::get('sites/count', [SitesController::class, 'count'])->name('app.sites.count');
+Route::post('sites', [SitesController::class, 'create'])
+    ->name('app.sites.add');
+Route::get('sites/sizes/{site_id?}', [SitesController::class, 'readSitesSizes'])
+    ->name('app.sites.sizes');
+Route::get('sites/{site}', [SitesController::class, 'read'])
+    ->name('app.sites.read');
+Route::patch('sites/{site}', [SitesController::class, 'update'])
+    ->name('app.sites.edit');
+Route::delete('sites/{site}', [SitesController::class, 'delete'])
+    ->name('app.sites.delete');
+Route::get('sites', [SitesController::class, 'list'])
+    ->name('app.sites.browse');
+Route::get('sites/count', [SitesController::class, 'count'])
+    ->name('app.sites.count');
 Route::put('sites/{site}/status', [SitesController::class, 'changeStatus'])
     ->name('app.sites.change_status');
 
@@ -77,9 +91,11 @@ Route::post('wallet/confirm-withdrawal', [WalletController::class, 'confirmWithd
 Route::delete('wallet/cancel-withdrawal/{entry}', [WalletController::class, 'cancelWithdrawal'])
     ->name('wallet.cancel-withdrawal');
 
-
 // statistics
-Route::get('campaigns/stats/chart/{type}/{resolution}/{date_start}/{date_end}', [StatsController::class, 'advertiserChart']);
+Route::get(
+    'campaigns/stats/chart/{type}/{resolution}/{date_start}/{date_end}',
+    [StatsController::class, 'advertiserChart']
+);
 Route::get('campaigns/stats/table/{date_start}/{date_end}', [StatsController::class, 'advertiserStats']);
 Route::get('campaigns/stats/table2/{date_start}/{date_end}', [StatsController::class, 'advertiserStatsWithTotal']);
 Route::get('sites/stats/chart/{type}/{resolution}/{date_start}/{date_end}', [StatsController::class, 'publisherChart']);