+++ src/Advertiser/Service/ChartDataProvider.php
@@ -29,12 +29,17 @@ use Adshares\Advertiser\Dto\Result\ChartResult;
 final class ChartDataProvider
 {
     private const REPOSITORY_MAPPER = [
-        StatsRepository::CLICK_TYPE => 'fetchClick',
-        StatsRepository::VIEW_TYPE => 'fetchView',
-        StatsRepository::CPC_TYPE => 'fetchCpc',
-        StatsRepository::CPM_TYPE => 'fetchCpm',
-        StatsRepository::SUM_TYPE => 'fetchSum',
-        StatsRepository::CTR_TYPE => 'fetchCtr',
+        StatsRepository::TYPE_CLICK => 'fetchClick',
+        StatsRepository::TYPE_CLICK_ALL => 'fetchClickAll',
+        StatsRepository::TYPE_CLICK_INVALID_RATE => 'fetchClickInvalidRate',
+        StatsRepository::TYPE_VIEW => 'fetchView',
+        StatsRepository::TYPE_VIEW_UNIQUE => 'fetchViewUnique',
+        StatsRepository::TYPE_VIEW_ALL => 'fetchViewAll',
+        StatsRepository::TYPE_VIEW_INVALID_RATE => 'fetchViewInvalidRate',
+        StatsRepository::TYPE_CPC => 'fetchCpc',
+        StatsRepository::TYPE_CPM => 'fetchCpm',
+        StatsRepository::TYPE_SUM => 'fetchSum',
+        StatsRepository::TYPE_CTR => 'fetchCtr',
     ];
 
     /** @var StatsRepository */